requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.40.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.12.2
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.1.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
zipp==3.16.0
